- Applied some refactorings
