+++ src/requests_rest_api/http_request_methods.py
@@ -1,6 +1,8 @@
 from enum import Enum
 from typing import Optional
 
+from http import HTTPStatus
+
 
 # supported request methods
 class RequestMethod(str, Enum):
@@ -14,12 +16,12 @@ class RequestMethod(str, Enum):
 
 # expected http status codes per method type
 STATUS_CODES_PER_REQUEST_METHOD = {
-    RequestMethod.GET: [200],
-    RequestMethod.HEAD: [200],
-    RequestMethod.POST: [200, 201, 204],
-    RequestMethod.PUT: [200, 202, 204],
-    RequestMethod.DELETE: [200, 202, 204],
-    RequestMethod.PATCH: [200, 204],
+    RequestMethod.GET: [HTTPStatus.OK],
+    RequestMethod.HEAD: [HTTPStatus.OK],
+    RequestMethod.POST: [HTTPStatus.OK, HTTPStatus.CREATED, HTTPStatus.NO_CONTENT],
+    RequestMethod.PUT: [HTTPStatus.OK, HTTPStatus.ACCEPTED, HTTPStatus.NO_CONTENT],
+    RequestMethod.DELETE: [HTTPStatus.OK, HTTPStatus.ACCEPTED, HTTPStatus.NO_CONTENT],
+    RequestMethod.PATCH: [HTTPStatus.OK, HTTPStatus.NO_CONTENT],
 }
